Firmware image creation Makefile and updater script
The updater script is a simpler alternative to the script used by
D-Link in their firmware updates, supporting only "userdata" updates.

Run "make" to create a new squashfs file system for the "userdata"
partition and pack, encrypt and sign it for use with the original
D-Link firmware update tools fwupdate or firmwareupgrade.cgi.

See the documetation in this repo for usage instructions.

Signed-off-by: Bjørn Mork <bjorn@mork.no>